support options.colorScheme
